requirements.txt
numpy
sentence-transformers
torch
fasttext-langdetect
spacy
asyncio
httpx
loguru
selectolax
pyopenssl
streamlit